requirements.txt
python-dotenv
fastapi
pydantic
uvicorn
Pillow
numpy
opencv-python-headless
mediapipe==0.10.9
inference_sdk